Refactor API into package
