+++ phone_spam_checker/api/jobs.py
@@ -1,81 +1,32 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Phone Checker API — Kaspersky / Truecaller / GetContact
-
-Run:
-    export API_KEY=supersecret
-    uvicorn phone_checker_api:app --host 0.0.0.0 --port 8000
-"""
-
 import asyncio
+import contextlib
+import logging
 import re
 import socket
-import contextlib
-from typing import List, Dict, Optional, Any
+from typing import Any, Dict, List, Optional
 
-import logging
-
-from fastapi import FastAPI, HTTPException, BackgroundTasks, Depends, Security
-from fastapi.security.api_key import APIKeyHeader
-from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
-import jwt
-from jwt import PyJWTError
+from fastapi import HTTPException
 from fastapi.responses import JSONResponse
-from pydantic import BaseModel, field_validator
 
+from phone_spam_checker.config import settings
+from phone_spam_checker.device_pool import DevicePool
+from phone_spam_checker.domain.models import PhoneCheckResult, CheckStatus
+from .schemas import CheckResult
+from phone_spam_checker.domain.phone_checker import PhoneChecker
+from phone_spam_checker.exceptions import DeviceConnectionError, JobAlreadyRunningError
 from phone_spam_checker.job_manager import (
     JobManager,
     SQLiteJobRepository,
     PostgresJobRepository,
 )
 from phone_spam_checker.logging_config import configure_logging
-from phone_spam_checker.config import settings
-from phone_spam_checker.device_pool import DevicePool
-
-# --- checker imports ----------------------------------------------------------
 from phone_spam_checker.registry import (
     get_checker_class,
     register_default_checkers,
     load_checker_module,
 )
-from phone_spam_checker.domain.models import PhoneCheckResult, CheckStatus
-from phone_spam_checker.domain.phone_checker import PhoneChecker
-from phone_spam_checker.exceptions import DeviceConnectionError, JobAlreadyRunningError
 
-# --- API key & token authorization -------------------------------------------
-API_KEY_NAME = "X-API-Key"
-api_key_header = APIKeyHeader(name=API_KEY_NAME, auto_error=False)
-bearer_scheme = HTTPBearer(auto_error=False)
-
-
-def _create_token() -> str:
-    return jwt.encode({"sub": "api"}, settings.secret_key, algorithm="HS256")
-
-
-async def login(api_key: str = Security(api_key_header)) -> dict:
-    if not settings.api_key or api_key != settings.api_key:
-        raise HTTPException(status_code=403, detail="Forbidden")
-    return {"access_token": _create_token()}
-
-
-async def get_token(
-    credentials: HTTPAuthorizationCredentials = Security(bearer_scheme),
-) -> str:
-    if credentials is None:
-        raise HTTPException(status_code=403, detail="Forbidden")
-    try:
-        jwt.decode(credentials.credentials, settings.secret_key, algorithms=["HS256"])
-    except PyJWTError:
-        raise HTTPException(status_code=403, detail="Invalid token")
-    return credentials.credentials
-
-
-configure_logging(
-    level=settings.log_level,
-    fmt=settings.log_format,
-    log_file=settings.log_file,
-)
+configure_logging(level=settings.log_level, fmt=settings.log_format, log_file=settings.log_file)
 register_default_checkers()
 for mod in filter(None, getattr(settings, "checker_modules", [])):
     load_checker_module(mod)
@@ -86,7 +37,10 @@ if settings.pg_host:
 else:
     job_repo = SQLiteJobRepository(settings.job_db_path)
 job_manager = JobManager(job_repo)
+
 job_queue: asyncio.Queue[tuple[str, List[str], str]] = asyncio.Queue()
+
+# device pools will be created on startup
 device_pools: Dict[str, DevicePool] = {}
 
 
@@ -122,67 +76,10 @@ async def _worker() -> None:
             job_queue.task_done()
 
 
-# --- data models (pydantic) ---------------------------------------------------
-from phone_spam_checker.validators import validate_phone_number
-
-
-class CheckRequest(BaseModel):
-    numbers: List[str]
-    service: str = "auto"  # 'auto', 'kaspersky', 'truecaller', 'getcontact'
-
-    @field_validator("numbers")
-    @classmethod
-    def validate_numbers(cls, v: List[str]) -> List[str]:
-        return [validate_phone_number(num) for num in v]
-
-
-class JobResponse(BaseModel):
-    job_id: str
-
-
-class CheckResult(BaseModel):
-    phone_number: str
-    status: CheckStatus
-    details: str = ""
-
-
-class StatusResponse(BaseModel):
-    job_id: str
-    status: str
-    results: Optional[List[CheckResult]] = None
-    error: Optional[str] = None
-
-
-# --- FastAPI ------------------------------------------------------------------
-app = FastAPI(
-    title="Phone Checker API",
-    version="2.0",
-)
-
-app.post("/login")(login)
-
-
-@app.exception_handler(DeviceConnectionError)
-async def device_error_handler(request, exc: DeviceConnectionError):
-    logger.error("Device connection error: %s", exc)
-    return JSONResponse(status_code=503, content={"detail": str(exc)})
-
-
-@app.middleware("http")
-async def exception_middleware(request, call_next):
-    try:
-        return await call_next(request)
-    except Exception as exc:
-        logger.exception("Unhandled exception")
-        raise HTTPException(status_code=500, detail="Internal server error") from exc
-
-
-# --- cleanup old jobs --------------------------------------------------------
 async def cleanup_jobs() -> None:
     await job_manager.cleanup_loop()
 
 
-@app.on_event("startup")
 async def start_background_tasks() -> None:
     asyncio.create_task(cleanup_jobs())
     device_pools.update(
@@ -198,9 +95,9 @@ async def start_background_tasks() -> None:
         asyncio.create_task(_worker())
 
 
-# ═════════════════════════════════════════════════════════════════════════════
-# 1. Kaspersky / Truecaller ----------------------------------------------------
-# ═════════════════════════════════════════════════════════════════════════════
+# =============================================================================
+# 1. Kaspersky / Truecaller
+# =============================================================================
 async def _run_check(
     job_id: str,
     numbers: List[str],
@@ -223,7 +120,7 @@ async def _run_check(
         kasp_nums = [n for n in numbers if re.match(r"^(7|\+7)9", n)]
         tc_nums = [n for n in numbers if n not in kasp_nums]
 
-    # -- ADB device addresses --------------------------------------------
+    # -- ADB device addresses
     kasp_device = kasp_device or f"{settings.kasp_adb_host}:{settings.kasp_adb_port}"
     tc_device = tc_device or f"{settings.tc_adb_host}:{settings.tc_adb_port}"
 
@@ -232,7 +129,7 @@ async def _run_check(
     loop = asyncio.get_event_loop()
 
     try:
-        # -- device initialization --------------------------------------
+        # -- device initialization
         if kasp_nums:
             _ping_device(*kasp_device.split(":"))
             kasp_checker_cls = get_checker_class("kaspersky")
@@ -247,7 +144,7 @@ async def _run_check(
             if not tc_checker.launch_app():
                 raise RuntimeError("Failed to launch Truecaller")
 
-        # -- parallel checking -------------------------------------------
+        # -- parallel checking
         tasks = []
         if kasp_nums:
             tasks.append(
@@ -264,7 +161,7 @@ async def _run_check(
 
         grouped = await asyncio.gather(*tasks)
 
-        # -- merge results ---------------------------------------------------
+        # -- merge results
         merged: Dict[str, Any] = {r.phone_number: r for group in grouped for r in group}
         for num in numbers:
             r = merged.get(num)
@@ -295,9 +192,9 @@ async def _run_check(
             await loop.run_in_executor(None, tc_checker.close_app)
 
 
-# ═════════════════════════════════════════════════════════════════════════════
-# 2. GetContact ---------------------------------------------------------------
-# ═════════════════════════════════════════════════════════════════════════════
+# =============================================================================
+# 2. GetContact
+# =============================================================================
 async def _run_check_gc(
     job_id: str, numbers: List[str], gc_device: Optional[str] = None
 ) -> None:
@@ -341,44 +238,8 @@ async def _run_check_gc(
             await loop.run_in_executor(None, checker.close_app)
 
 
-# --- endpoints ---------------------------------------------------------------
-@app.post("/check_numbers", response_model=JobResponse)
-def submit_check(
-    request: CheckRequest,
-    background_tasks: BackgroundTasks,
-    _: str = Depends(get_token),
-) -> JobResponse:
-    try:
-        job_id = _new_job(request.service)
-    except JobAlreadyRunningError as e:
-        raise HTTPException(status_code=429, detail=str(e)) from e
-    logger.info(
-        "Received job %s: %d numbers via %s",
-        job_id,
-        len(request.numbers),
-        request.service,
-    )
-    if request.service not in {"auto", "kaspersky", "truecaller", "getcontact"}:
-        raise HTTPException(status_code=400, detail="Unknown service")
+# Helper functions
 
-    background_tasks.add_task(enqueue_job, job_id, request.numbers, request.service)
-    return JobResponse(job_id=job_id)
-
-
-@app.get("/status/{job_id}", response_model=StatusResponse)
-def get_status(job_id: str, _: str = Depends(get_token)) -> StatusResponse:
-    job = job_manager.get_job(job_id)
-    if not job:
-        raise HTTPException(status_code=404, detail="Job ID not found")
-    return StatusResponse(
-        job_id=job_id,
-        status=job["status"],
-        results=job.get("results"),
-        error=job.get("error"),
-    )
-
-
-# --- helper functions --------------------------------------------------------
 def _ping_device(host: str, port: str, timeout: int = 5) -> None:
     logger.debug("Pinging device %s:%s", host, port)
     try:
@@ -407,6 +268,7 @@ def _ensure_no_running(service: str) -> None:
     except JobAlreadyRunningError as e:
         raise HTTPException(status_code=429, detail=str(e)) from e
 
+
 def _new_job(service: str) -> str:
     job_id = job_manager.new_job(_devices_for_service(service))
     logger.debug("Created job %s", job_id)
@@ -421,3 +283,16 @@ def _complete_job(job_id: str, results: List[CheckResult]) -> None:
 def _fail_job(job_id: str, error: str) -> None:
     logger.debug("Marking job %s as failed: %s", job_id, error)
     job_manager.fail_job(job_id, error)
+
+
+def device_error_handler(request, exc: DeviceConnectionError):
+    logger.error("Device connection error: %s", exc)
+    return JSONResponse(status_code=503, content={"detail": str(exc)})
+
+
+async def exception_middleware(request, call_next):
+    try:
+        return await call_next(request)
+    except Exception as exc:
+        logger.exception("Unhandled exception")
+        raise HTTPException(status_code=500, detail="Internal server error") from exc
